agent/pyproject.toml
[project]
name = "myagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "openai>=2.3.0",
    "openai-agents>=0.1.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
